# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Platform for online internships",
    "version": "v1"
  },
  "paths": {
    "/api/courses/": {
      "get": {
        "summary": "Get Courses",
        "operationId": "get_courses_api_courses__get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_Course_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Course",
        "operationId": "create_course_api_courses__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__api__schemas__courses__CreateCourse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneCourse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/courses/{course_id}": {
      "get": {
        "summary": "Get Course",
        "operationId": "get_course_api_courses__course_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneCourse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Course",
        "operationId": "delete_course_api_courses__course_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Patch Course",
        "operationId": "patch_course_api_courses__course_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCourse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneCourse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/user/{user_id}/courses/": {
      "get": {
        "summary": "Get User Courses",
        "operationId": "get_user_courses_api_user__user_id__courses__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            },
            "name": "user_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_NamedUserCourse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create User Course",
        "operationId": "create_user_course_api_user__user_id__courses__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__api__schemas__user_courses__CreateCourse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCourse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/courses/{course_id}": {
      "get": {
        "summary": "Get One User Course",
        "operationId": "get_one_user_course_api_user__user_id__courses__course_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneUserCourse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete User Course",
        "operationId": "delete_user_course_api_user__user_id__courses__course_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/competencies/": {
      "get": {
        "summary": "Get Competencies",
        "operationId": "get_competencies_api_competencies__get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_Competence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_competencies__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/competencies/{competence_id}": {
      "get": {
        "summary": "Get Competence",
        "operationId": "get_competence_api_competencies__competence_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_competencies__competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Patch Competence",
        "operationId": "patch_competence_api_competencies__competence_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/competencies/": {
      "get": {
        "summary": "Get Course Competencies",
        "operationId": "get_course_competencies_api_courses__course_id__competencies__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_CourseCompetence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_courses__course_id__competencies__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCompetence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/competencies/{competence_id}/courses/": {
      "get": {
        "summary": "Get Related Courses",
        "operationId": "get_related_courses_api_competencies__competence_id__courses__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Competence Id",
              "type": "integer"
            },
            "name": "competence_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_Course_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/competencies/{course_competence_id}": {
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_courses__course_id__competencies__course_competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Course Competence Id",
              "type": "integer"
            },
            "name": "course_competence_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            },
            "name": "course_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/subdivisions/{subdivision_id}/posts/{post_id}/competencies/": {
      "get": {
        "summary": "Get Post Competencies",
        "operationId": "get_post_competencies_api_subdivisions__subdivision_id__posts__post_id__competencies__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "minimum": 1.0,
              "type": "integer",
              "default": 20
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Search",
              "type": "string"
            },
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_PostCompetence_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Competence",
        "operationId": "create_competence_api_subdivisions__subdivision_id__posts__post_id__competencies__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostCompetence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCompetence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/subdivisions/{subdivision_id}/posts/{post_id}/competencies/{post_competence_id}": {
      "delete": {
        "summary": "Delete Compentence",
        "operationId": "delete_compentence_api_subdivisions__subdivision_id__posts__post_id__competencies__post_competence_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Post Competence Id",
              "type": "integer"
            },
            "name": "post_competence_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Post Id"
            },
            "name": "post_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateCompetence": {
        "title": "CreateCompetence",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "LimitOffsetPage_Course_": {
        "title": "LimitOffsetPage[Course]",
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1.0,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0.0,
            "type": "integer"
          }
        }
      },
...
      "LimitOffsetPage_NamedUserCourse_": {
        "title": "LimitOffsetPage[NamedUserCourse]",
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedUserCourse"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1.0,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0.0,
            "type": "integer"
          }
        }
      },
...
      "OneCourse": {
        "title": "OneCourse",
        "required": [
          "id",
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        }
      },
...
      "PatchCourse": {
        "title": "PatchCourse",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        }
      },
...
      "UserCourse": {
        "title": "UserCourse",
        "required": [
          "id",
          "user_id",
          "course_id",
          "progress",
          "admission_date"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "admission_date": {
            "title": "Admission Date",
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "backend__api__schemas__courses__CreateCourse": {
        "title": "CreateCourse",
        "required": [
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        }
      },
      "backend__api__schemas__user_courses__CreateCourse": {
        "title": "CreateCourse",
        "required": [
          "course_id"
        ],
        "type": "object",
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}